+++ src/app/views/admin/pages/others/espacepointfocalcom/espacepointfocalcom.component.ts
@@ -425,7 +425,7 @@ export class EspacepointfocalcomComponent implements OnInit {
       AppSweetAlert.simpleAlert("Erreur", "Vous ne pouvez plus supprimer cette requête. Elle est déjà en cours de traitement.", 'error');
       return;
     }
-    AlertNotif.finishConfirm("Suppression requete",
+    AppSweetAlert.confirmBox("Suppression requete",
       "Cette action est irreversible. Voulez-vous continuer ?").then((result:any) => {
         if (result.value) {
           this.requeteService.delete(this.selected_data.id).subscribe((res: any) => {
@@ -464,7 +464,7 @@ export class EspacepointfocalcomComponent implements OnInit {
       AppSweetAlert.simpleAlert("Erreur", "Votre de demande est déjà en cours de traitement.", 'error');
       return;
     }
-    AlertNotif.finishConfirm("Transmettre rdv",
+    AppSweetAlert.confirmBox("Transmettre rdv",
       "Cette action est irreversible. Voulez-vous continuer ?").then((result) => {
         if (result.value) {
           var param = {
@@ -492,7 +492,7 @@ export class EspacepointfocalcomComponent implements OnInit {
       AppSweetAlert.simpleAlert("Erreur", "Vous ne pouvez plus supprimer cet element. Elle est déjà en cours de traitement.", 'error');
       return;
     }
-    AlertNotif.finishConfirm("Suppression rdv",
+    AppSweetAlert.confirmBox("Suppression rdv",
       "Cette action est irreversible. Voulez-vous continuer ?").then((result:any) => {
         if (result.value) {
           this.rdvService.delete(this.selected_data2.id).subscribe((res: any) => {
