+++ src/app/views/admin/pages/others/list-requete-usager/list-requete-usager.component.ts
@@ -394,7 +394,7 @@ export class ListRequeteUsagerComponent implements OnInit {
       AppSweetAlert.simpleAlert("Erreur", "Vous ne pouvez plus supprimer cette requête. Elle est déjà en cours de traitement.", 'error');
       return;
     }
-    AlertNotif.finishConfirm("Suppression requete",
+    AppSweetAlert.confirmBox("Suppression requete",
       "Cette action est irreversible. Voulez-vous continuer ?").then((result:any) => {
         if (result.value) {
           this.requeteService.delete(this.selected_data.id).subscribe((res: any) => {
@@ -573,7 +573,7 @@ export class ListRequeteUsagerComponent implements OnInit {
       idEntiteReceive: value.idEntiteReceive,
       idPrestation: value.idPrestation,
     };
-    AlertNotif.finishConfirm("Transférer cette préoccupation à un autre ministère/institution",
+    AppSweetAlert.confirmBox("Transférer cette préoccupation à un autre ministère/institution",
     "Cette action est irreversible. Voulez-vous continuer ?").then((result:any) => {
       if (result.value) {
         this.requeteService.transfertRequet(param,this.selected_data.id).subscribe((rest: any) => {
@@ -602,7 +602,7 @@ export class ListRequeteUsagerComponent implements OnInit {
       idEntite: this.user.idEntite,
       idPrestation: value.idPrestation,
     };
-    AlertNotif.finishConfirm("Transférer cette préoccupation à la structure",
+    AppSweetAlert.confirmBox("Transférer cette préoccupation à la structure",
     "Cette action est irreversible. Voulez-vous continuer ?").then((result:any) => {
       if (result.value) {
         this.requeteService.transfertRequetInterne(param,this.selected_data.id).subscribe((rest: any) => {
@@ -636,7 +636,7 @@ export class ListRequeteUsagerComponent implements OnInit {
       idEntite: this.user.idEntite,
       idPrestation: value.idPrestation,
     };
-    AlertNotif.finishConfirm("Réorienter cette préoccupation",
+    AppSweetAlert.confirmBox("Réorienter cette préoccupation",
     "Cette action est irreversible. Voulez-vous continuer ?").then((result:any) => {
       if (result.value) {
         this.requeteService.transfertRequetInterne(param,this.selected_data.id).subscribe((rest: any) => {
