+++ src/app/views/admin/pages/others/manage-requete-usager/manage-requete-usager.component.ts
@@ -385,7 +385,7 @@ export class ManageRequeteUsagerComponent implements OnInit {
       AppSweetAlert.simpleAlert("Erreur", "Vous ne pouvez plus supprimer cette requête. Elle est déjà en cours de traitement.", 'error');
       return;
     }
-    AlertNotif.finishConfirm("Suppression requete",
+    AppSweetAlert.confirmBox("Suppression requete",
       "Cette action est irreversible. Voulez-vous continuer ?").then((result:any) => {
         if (result.value) {
           this.requeteService.delete(this.selected_data.id).subscribe((res: any) => {
@@ -549,7 +549,7 @@ export class ManageRequeteUsagerComponent implements OnInit {
       idEntiteReceive: value.idEntiteReceive,
       idPrestation: value.idPrestation,
     };
-    AlertNotif.finishConfirm("Transférer cette préoccupation à un autre ministère/institution",
+    AppSweetAlert.confirmBox("Transférer cette préoccupation à un autre ministère/institution",
     "Cette action est irreversible. Voulez-vous continuer ?").then((result:any) => {
       if (result.value) {
         this.requeteService.transfertRequet(param,this.selected_data.id).subscribe((rest: any) => {
